+++ src/ipm/ipx/src/lp_solver.h
@@ -147,6 +147,24 @@ public:
     // Returns -1 if no basis was available and 0 otherwise.
     Int SymbolicInvert(Int* rowcounts, Int* colcounts);
 
+    void RunCrossover_() {
+        iterate_.reset(new Iterate(model_));
+
+        BuildStartingBasis();
+
+        RunCrossover();
+    }
+
+    void RunCrossover_X() {
+        iterate_.reset(new Iterate(model_));
+
+        iterate_->Initialize(x_start_, xl_start_, xu_start_,
+                             y_start_, zl_start_, zu_start_);
+        BuildStartingBasis();
+
+        RunCrossover();
+    }
+
 private:
     void ClearSolution();
     void InteriorPointSolve();
